+++ src/commands/config/add_connection/mod.rs
@@ -116,19 +116,19 @@ impl AddNetworkConnection {
     ) -> color_eyre::eyre::Result<Option<crate::types::api_key::ApiKey>> {
         #[derive(strum_macros::Display)]
         enum ConfirmOptions {
-            #[strum(to_string = "Yes, the RPC endpoint requires API key")]
+            #[strum(to_string = "Yes, the RPC endpoint requires an API key")]
             Yes,
-            #[strum(to_string = "No, the RPC endpoint does not require API key")]
+            #[strum(to_string = "No, the RPC endpoint does not require an API key")]
             No,
         }
         let select_choose_input = Select::new(
-            "Do you want to input an API key?",
+            "Do you want to enter an API key?",
             vec![ConfirmOptions::Yes, ConfirmOptions::No],
         )
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let api_key: crate::types::api_key::ApiKey =
-                CustomType::new("Enter an API key").prompt()?;
+                CustomType::new("Enter an API key:").prompt()?;
             Ok(Some(api_key))
         } else {
             Ok(None)
@@ -144,7 +144,7 @@ impl AddNetworkConnection {
                 to_string = "Yes, and I want to enter the name of the account hosting the program \"linkdrop\""
             )]
             Yes,
-            #[strum(to_string = "I dont know")]
+            #[strum(to_string = "I don't know")]
             No,
         }
         let select_choose_input = Select::new(
@@ -154,7 +154,7 @@ impl AddNetworkConnection {
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let account_id: crate::types::account_id::AccountId =
-            CustomType::new("What is the name of the account that hosts the \"linkdrop\" program? (e.g. on mainnet it is near, and on testnet it is testnet)").prompt()?;
+            CustomType::new("What is the name of the account that hosts the linkdrop program? (e.g. on mainnet it is 'near', on testnet it is 'testnet')").prompt()?;
             Ok(Some(account_id))
         } else {
             Ok(None)
@@ -180,7 +180,7 @@ impl AddNetworkConnection {
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let account_id: crate::types::account_id::AccountId =
-            CustomType::new("What is the name of the NEAR Social DB contract account ID (e.g. on mainnet it is social.near)").prompt()?;
+            CustomType::new("What is the NEAR Social DB contract account ID? (e.g. on mainnet it is 'social.near')").prompt()?;
             Ok(Some(account_id))
         } else {
             Ok(None)
@@ -204,7 +204,7 @@ impl AddNetworkConnection {
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let faucet_url: crate::types::url::Url =
-                CustomType::new("What is the faucet url?").prompt()?;
+                CustomType::new("What is the faucet URL?").prompt()?;
             Ok(Some(faucet_url))
         } else {
             Ok(None)
@@ -228,7 +228,7 @@ impl AddNetworkConnection {
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let meta_transaction_relayer_url: crate::types::url::Url =
-                CustomType::new("What is the relayer url?").prompt()?;
+                CustomType::new("What is the meta transaction relayer URL?").prompt()?;
             Ok(Some(meta_transaction_relayer_url))
         } else {
             Ok(None)
@@ -240,19 +240,19 @@ impl AddNetworkConnection {
     ) -> color_eyre::eyre::Result<Option<crate::types::url::Url>> {
         #[derive(strum_macros::Display)]
         enum ConfirmOptions {
-            #[strum(to_string = "Yes, I want to enter the fastnear API url")]
+            #[strum(to_string = "Yes, I want to enter the FastNEAR API URL")]
             Yes,
-            #[strum(to_string = "No, I don't want to enter the fastnear API url")]
+            #[strum(to_string = "No, I don't want to enter the FastNEAR API URL")]
             No,
         }
         let select_choose_input = Select::new(
-            "Do you want to enter the fastnear API url?",
+            "Do you want to enter the FastNEAR API URL?",
             vec![ConfirmOptions::Yes, ConfirmOptions::No],
         )
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let stake_delegators_api: crate::types::url::Url =
-                CustomType::new("What is the fastnear API url?").prompt()?;
+                CustomType::new("What is the FastNEAR API URL?").prompt()?;
             Ok(Some(stake_delegators_api))
         } else {
             Ok(None)
@@ -288,19 +288,19 @@ impl AddNetworkConnection {
     ) -> color_eyre::eyre::Result<Option<crate::types::url::Url>> {
         #[derive(strum_macros::Display)]
         enum ConfirmOptions {
-            #[strum(to_string = "Yes, I want to enter the coingecko API url")]
+            #[strum(to_string = "Yes, I want to enter the CoinGecko API URL")]
             Yes,
-            #[strum(to_string = "No, I don't want to enter the coingecko API url")]
+            #[strum(to_string = "No, I don't want to enter the CoinGecko API URL")]
             No,
         }
         let select_choose_input = Select::new(
-            "Do you want to enter the coingecko API url?",
+            "Do you want to enter the CoinGecko API URL?",
             vec![ConfirmOptions::Yes, ConfirmOptions::No],
         )
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let coingecko_api: crate::types::url::Url =
-                CustomType::new("What is the coingecko API url?")
+                CustomType::new("What is the CoinGecko API URL?")
                     .with_starting_input("https://api.coingecko.com/")
                     .prompt()?;
             Ok(Some(coingecko_api))
@@ -326,7 +326,8 @@ impl AddNetworkConnection {
         .prompt()?;
         if let ConfirmOptions::Yes = select_choose_input {
             let mpc_contract_account_id: crate::types::account_id::AccountId =
-                CustomType::new("What is the MPC contract account ID?").prompt()?;
+                CustomType::new("What is the MPC (Multi-Party Computation) contract account ID?")
+                    .prompt()?;
             Ok(Some(mpc_contract_account_id))
         } else {
             Ok(None)