+++ src/commands/staking/delegate/deposit_and_stake.rs
@@ -2,7 +2,7 @@
 #[interactive_clap(input_context = super::StakeDelegationContext)]
 #[interactive_clap(output_context = DepositAndStakeContext)]
 pub struct DepositAndStake {
-    /// Enter the attached amount to be deposited and then staked into the predecessor's internal account (example: 10NEAR or 0.5near or 10000yoctonear):
+    /// Enter the attached amount to be deposited and then staked into the predecessor's internal account (example: 10 NEAR or 0.5 NEAR or 10000 yoctonear):
     amount: crate::types::near_token::NearToken,
     #[interactive_clap(skip_default_input_arg)]
     /// What is validator account ID?