+++ src/transaction_signature_options/sign_with_ledger/mod.rs
@@ -388,7 +388,7 @@ impl SignLedger {
 pub fn input_seed_phrase_hd_path(
 ) -> color_eyre::eyre::Result<Option<crate::types::slip10::BIP32Path>> {
     Ok(Some(
-        CustomType::new("Enter seed phrase HD Path (if you not sure leave blank for default):")
+        CustomType::new("Enter seed phrase HD Path (if you're not sure, leave blank for default):")
             .with_starting_input("44'/397'/0'/0'/1'")
             .prompt()?,
     ))